refactor(global-state): restructured global state
