+++ frontend/src/store/last-will/LastWillContext.tsx
@@ -2,17 +2,7 @@
 import { createContext, useCallback, useContext, useEffect, useReducer } from 'react'
 import { initalLastWillState, lastWillReducer } from './reducer'
 import { submitTestatorAction } from './testator/actions'
-import { TesatatorFormPayload } from './testator/payload'
-import { LastWill } from './type'
-
-type LastWillContextType = {
-	lastWill: LastWill
-	services: {
-		submitTestator: (payload: TesatatorFormPayload) => void
-	}
-}
-
-const LastWillContext = createContext({} as LastWillContextType)
+import { LastWillContextType } from './types'
 
 /**
  * Handles the Testament object where we hold all data.
@@ -20,20 +10,18 @@ const LastWillContext = createContext({} as LastWillContextType)
 export const LastWillContextProvider: React.FC<{ children: React.ReactNode }> = ({ children }) => {
 	const [lastWill, dispatch] = useReducer(lastWillReducer, initalLastWillState)
 
-	/**
-	 * Reload fetches the data from the backend and updates the context.
-	 */
 	const reloadData = useCallback(async () => {}, [])
-
-	const submitTestator = useCallback<LastWillContextType['services']['submitTestator']>(
-		(payload) => submitTestatorAction(dispatch, payload),
-		[]
-	)
 
 	// Reload data on mount
 	useEffect(() => {
 		reloadData()
 	}, [reloadData])
+
+	// Global State Services 
+	const submitTestator = useCallback<LastWillContextType['services']['submitTestator']>(
+		(payload) => submitTestatorAction(dispatch, payload),
+		[]
+	)
 
 	return (
 		<LastWillContext.Provider
@@ -49,4 +37,5 @@ export const LastWillContextProvider: React.FC<{ children: React.ReactNode }> = 
 	)
 }
 
+const LastWillContext = createContext({} as LastWillContextType)
 export const useLastWillContext = () => useContext(LastWillContext)
