+++ frontend/src/store/last-will/testator/api.ts
@@ -1,6 +1,11 @@
-import { TesatatorFormPayload, TestatorResponse } from './payload'
+export type TestatorFormPayload = {
+	name: string
+}
+export type TestatorResponse = {
+	name: string
+}
 
-export const saveTestator = async (payload: TesatatorFormPayload) => {
+export const saveTestator = async (payload: TestatorFormPayload) => {
 	// simulate async request
 	const data = await new Promise<TestatorResponse>((resolve) => {
 		setTimeout(() => {
