--- frontend/src/store/last-will/testator/payload.ts
@@ -1,6 +1,0 @@
-export type TesatatorFormPayload = {
-	name: string
-}
-export type TestatorResponse = {
-	name: string
-}
