--- frontend/src/store/last-will/type.ts
@@ -1,9 +1,0 @@
-import { TestatorState } from './testator/state'
-
-export type LastWill = {
-	testator: TestatorState
-	marriageStatus: string
-	heirs: string
-	inheritance: string
-	succession: string
-}
